Ajout des gestions des lieux, ajout de leaflet, correction de l'envoi des mails

+++ src/Controller/CampusController.php
@@ -5,6 +5,8 @@ namespace App\Controller;
 use App\Entity\Campus;
 use App\Form\CampusType;
 use App\Repository\CampusRepository;
+use App\Repository\ParticipantRepository;
+use App\Repository\SortieRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -63,11 +65,19 @@ final class CampusController extends AbstractController
     }
 
     #[Route(path: '/{id}', name: 'app_campus_delete', methods: ['POST'])]
-    public function delete(Request $request, Campus $campus, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, Campus $campus, EntityManagerInterface $entityManager, SortieRepository $sortieRepository, ParticipantRepository $participantRepository): Response
     {
         if ($this->isCsrfTokenValid('delete'.$campus->getId(), $request->getPayload()->getString('_token'))) {
-            $entityManager->remove($campus);
-            $entityManager->flush();
+            $sorties = $sortieRepository->findBy(['campus' => $campus]);
+            $participants = $participantRepository->findBy(['campus' => $campus]);
+
+            if (count($sorties) > 0 || count($participants) > 0) {
+                $this->addFlash('danger', 'Impossible de supprimer le campus car il est utilisé par des sorties ou des participants.');
+            } else {
+                $entityManager->remove($campus);
+                $entityManager->flush();
+                $this->addFlash('success', 'Le campus a été supprimé avec succès.');
+            }
         }
 
         return $this->redirectToRoute('app_campus_index', [], Response::HTTP_SEE_OTHER);
